reconstruction tests restructured and updated, init_vol bug
