+++ savu/test/travis/plugin_tests/reconstruction_tests/astra/__init__.py
@@ -13,27 +13,9 @@
 # limitations under the License.
 
 """
-.. module:: tomo_recon
-   :platform: Unix
-   :synopsis: runner for tests using the MPI framework
+Test suites for Savu are here
+
 
 .. moduleauthor:: Mark Basham <scientificsoftware@diamond.ac.uk>
 
 """
-import unittest
-from savu.test import test_utils as tu
-from savu.test.travis.framework_tests.plugin_runner_test \
-    import run_protected_plugin_runner
-
-
-class SimpleTomoTest(unittest.TestCase):
-
-    def test_process(self):
-        data_file = tu.get_test_data_path('24737.nxs')
-        process_file = \
-            tu.get_test_process_path('simple_recon_test_process.nxs')
-        run_protected_plugin_runner(tu.set_options(data_file,
-                                                   process_file=process_file))
-
-if __name__ == "__main__":
-    unittest.main()
