+++ savu/test/travis/plugin_tests/reconstruction_tests/simple_recon/fbp_test.py
@@ -23,18 +23,33 @@
 """
 
 import unittest
-
 import savu.test.test_utils as tu
 from savu.test.travis.framework_tests.plugin_runner_test import \
-    run_protected_plugin_runner_no_process_list
+        run_protected_plugin_runner
+import savu.test.base_checkpoint_test
+import tempfile
+import os
 
 
-class SimpleReconTest(unittest.TestCase):
+class FbpTest(unittest.TestCase):
 
-    def test_simple_recon(self):
+    def test_fbp(self):
+        data_file = tu.get_test_data_path('24737.nxs')
+        self.test_folder = tempfile.mkdtemp(suffix='my_test/')
+        # set options
         options = tu.set_experiment('tomo')
-        plugin = 'savu.plugins.reconstructions.simple_recon'
-        run_protected_plugin_runner_no_process_list(options, plugin)
+        options['data_file'] = data_file
+        options['out_path'] = os.path.join(self.test_folder)
+        options['process_file'] = tu.get_test_process_path('fbp/simple_recon_test_process.nxs')
+        run_protected_plugin_runner(options)
+
+        # perform folder cleaning
+        classb = savu.test.base_checkpoint_test.BaseCheckpointTest()
+        cp_folder = os.path.join(self.test_folder, 'checkpoint')
+        classb._empty_folder(cp_folder)
+        os.removedirs(cp_folder)
+        classb._empty_folder(self.test_folder)
+        os.removedirs(self.test_folder)
 
 #
 #class ScikitimageSartTest(unittest.TestCase):
